rag_ai_agent_tutorials/multimodal_agentic_rag/backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
google-genai>=1.0.0
google-adk>=1.0.0
python-multipart>=0.0.7
beautifulsoup4>=4.12.0
httpx>=0.27.0
